% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alert_of_drop}
\alias{alert_of_drop}
\title{Title}
\usage{
alert_of_drop(reactions_to_drop, cause, is_warning = FALSE)
}
\arguments{
\item{reactions_to_drop}{the first param}

\item{cause}{the second param}

\item{is_warning}{the third param}
}
\value{
the output
}
\description{
Description
}
