% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyzer.R
\name{Analyzer}
\alias{Analyzer}
\title{Title}
\description{
Description.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{A field.}

\item{\code{reaction_consistencies}}{A field.}

\item{\code{metareaction_consistencies}}{A field.}

\item{\code{gene_expression_statistics}}{A field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Analyzer$new()}}
\item \href{#method-print}{\code{Analyzer$print()}}
\item \href{#method-repr}{\code{Analyzer$repr()}}
\item \href{#method-clone}{\code{Analyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyzer$new(
  ...,
  metadata_directory,
  gene_metadata_file,
  metabolite_metadata_file,
  reaction_metadata_file,
  reaction_consistencies_file,
  linear_gene_expression_matrix_file,
  reaction_annotation_separator,
  reaction_annotations,
  min_reaction_consistency = 1e-04,
  min_reaction_range = 1e-08,
  cluster_strength = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}

\item{\code{metadata_directory}}{A param.}

\item{\code{gene_metadata_file}}{A param.}

\item{\code{metabolite_metadata_file}}{A param.}

\item{\code{reaction_metadata_file}}{A param.}

\item{\code{reaction_consistencies_file}}{A param.}

\item{\code{linear_gene_expression_matrix_file}}{A param.}

\item{\code{reaction_annotation_separator}}{A param.}

\item{\code{reaction_annotations}}{A param.}

\item{\code{min_reaction_consistency}}{A param.}

\item{\code{min_reaction_range}}{A param.}

\item{\code{cluster_strength}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyzer$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\subsection{Method \code{repr()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyzer$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
