% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster.R
\name{avg_sil}
\alias{avg_sil}
\title{avg_sil}
\usage{
avg_sil(df, k, method = "average")
}
\arguments{
\item{df}{a matrix or data.frame with coordinates.}

\item{k}{the desired number of clusters.}

\item{method}{the agglomeration method to be used.  This should
    be (an unambiguous abbreviation of) one of
    \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"},
    \code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA),
    \code{"median"} (= WPGMC) or \code{"centroid"} (= UPGMC).}
}
\value{
the average silhouette index.
}
\description{
Given a matrix or data.frame with point coordinates and the number of clusters to
get, the average silhouette is computed.
}
\examples{
set.seed(123)
mat <- matrix(c(
    rnorm(n = 100, mean = 10),
    rnorm(n = 100, mean = 20),
    rnorm(n = 100, mean = 1),
    rnorm(n = 100, mean = 2)), ncol = 2)
avg_sil(mat, 2)
}
