% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gene}
\alias{check_gene}
\title{check if a gene is valid}
\usage{
check_gene(assembly, gene)
}
\arguments{
\item{assembly}{hg38 or mm10}

\item{gene}{gene symbol}
}
\value{
whether the gene is valid
}
\description{
check if a gene is valid
}
\examples{
NA
}
