% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{combind_plots}
\alias{combind_plots}
\title{plot_expr_boxplot = function(expr_list_all, group_df, gene, p_group){
  sample_ids = group_df$Sample
  expr_list = lapply(sample_ids, function(x){
    expr_df = data.frame(expr_list_all[[x]][gene, ])
    expr_df$Sample = x
    colnames(expr_df) = c("Gene", "Sample")
    return(expr_df)
  })
  plot_df = do.call("rbind", expr_list)
  plot_df$Sample = factor(plot_df$Sample, levels = rev(sample_ids))
  p <- ggplot(data=plot_df, aes(x=Sample, y=Gene)) +
    labs(y = gene) +
    # geom_bar(stat="identity") +
    geom_boxplot(fill='#A4A4A4') +
    scale_y_continuous(position = "right", expand = c(0,0)) +
    coord_flip() +
    ylim2(p_group) +
    theme(
      panel.background = element_rect(fill='transparent'),
      plot.background = element_rect(fill='transparent', color=NA),
      axis.text.y=element_blank(),
      axis.title.y=element_blank()
    )
  return(p)
}
Merge all plots together a large plot}
\usage{
combind_plots(
  track_plot,
  annot_plot,
  expr_plot,
  track_width,
  legend.position = "bottom",
  t = -20,
  b = 20
)
}
\arguments{
\item{track_plot}{plot object of genome tracks}

\item{annot_plot}{annotation bar for expression}

\item{expr_plot}{gene expression barplot}

\item{track_width}{width of tracks}

\item{legend.position}{legend position, by default will be bottom}

\item{t}{for alignment of expression barplot, distance to top}

\item{b}{for alignment of expression barplot, distance to bottom}
}
\value{
plot object of combined plot
}
\description{
plot_expr_boxplot = function(expr_list_all, group_df, gene, p_group){
  sample_ids = group_df$Sample
  expr_list = lapply(sample_ids, function(x){
    expr_df = data.frame(expr_list_all[[x]][gene, ])
    expr_df$Sample = x
    colnames(expr_df) = c("Gene", "Sample")
    return(expr_df)
  })
  plot_df = do.call("rbind", expr_list)
  plot_df$Sample = factor(plot_df$Sample, levels = rev(sample_ids))
  p <- ggplot(data=plot_df, aes(x=Sample, y=Gene)) +
    labs(y = gene) +
    # geom_bar(stat="identity") +
    geom_boxplot(fill='#A4A4A4') +
    scale_y_continuous(position = "right", expand = c(0,0)) +
    coord_flip() +
    ylim2(p_group) +
    theme(
      panel.background = element_rect(fill='transparent'),
      plot.background = element_rect(fill='transparent', color=NA),
      axis.text.y=element_blank(),
      axis.title.y=element_blank()
    )
  return(p)
}
Merge all plots together a large plot
}
\examples{
NA
}
