% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{compare_clusters}
\alias{compare_clusters}
\title{compare_clusters}
\usage{
compare_clusters(
  compass_analyzer,
  compass_data,
  variable,
  cluster_A,
  cluster_B,
  for_metareactions = FALSE,
  adjusted_p_value_th = 0.05,
  cohens_d_th = 1
)
}
\arguments{
\item{compass_analyzer}{the object created by CompassAnalyzer$new}

\item{compass_data}{the object created by CompassData$new}

\item{variable}{the name of the variable to test}

\item{cluster_A, cluster_B}{vector of labels for the chosen variable levels 
to compare}

\item{for_metareactions}{logical to decide wether to perform tests for 
reactions or metareactions}

\item{adjusted_p_value_th}{adjusted p-value threshold to detect significant 
reactions (default 0.05)}

\item{cohens_d_th}{cohen's D statistic threshold to detect effect-size 
significant reactions (default 1)}
}
\value{
a list of results.
\itemize{
    \item{\code{wilcoxon_results_with_metadata}}{ wilcoxon test results}
    \item{\code{DC_core_metabolism}}{ table of differentially consistent 
    reactions or metareactions}
    \item{\code{DC_core_metabolism_table}}{ ready-to-plot table of
    differentially consistent reactions or metareactions}
    \item{\code{p_DC_reactions}}{ plot for the differentially consistent 
    reactions when \code{for_metareactions = TRUE}}
    \item{\code{p_DC_reactions_cohensd}}{ plot for the effect sizes of the 
    differentially consistent reactions when \code{for_metareactions = TRUE}}}
}
\description{
Given two groups of cells the function returns several objects to study the 
differentially consistent reactions.
}
