% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompassSettings.R
\name{CompassSettings}
\alias{CompassSettings}
\title{CompassSettings}
\description{
An object through which you can specify settings for your COMPASS analysis.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gene_metadata_path}}{A path to a file containing tabular gene metadata. Each row should represent a single gene. The columns are up to you, so long as one of them provides a unique identifier for each gene.}

\item{\code{metabolite_metadata_path}}{A path to a file containing tabular metabolite metadata. Each row should represent a single metabolite. The columns are up to you, so long as one of them provides a unique identifier for each metabolite.}

\item{\code{reaction_metadata_path}}{A path to a file containing tabular reaction metadata. Each row should represent a single reaction. The columns are up to you, so long as one of them provides a unique identifier for each reaction.}

\item{\code{cell_metadata_path}}{A path to a file containing tabular cell metadata. Each row should represent a single cell. The columns are up to you, so long as one of them provides a unique identifier for each cell.}

\item{\code{compass_scores_path}}{A path to a file containing the raw reaction consistencies matrix. (This is the output of the COMPASS algorithm.)}

\item{\code{linear_gene_expression_matrix_path}}{A path to a file containing the linear gene expression matrix. (This is the input to the COMPASS algorithm.)}

\item{\code{cell_id_col_name}}{The name of the column that uniquely identifies each cell in the cell metadata file.}

\item{\code{gene_id_col_name}}{The name of the column that uniquely identifies each gene in the gene metadata file.}

\item{\code{reaction_direction_separator}}{It is assumed that reaction IDs take the form {unique id}{separator}{one of N annotations}, where the separator is specified by this length-1 character vector, interpreted as a regular expression.}

\item{\code{reaction_directions}}{It is assumed that reaction IDs take the form {unique id}{separator}{one of N annotations}, where the annotations are specified by this length-N character vector, interpreted as regular expressions.}

\item{\code{min_reaction_consistency}}{Reactions are dropped that have consistency scores below this threshold.}

\item{\code{min_reaction_range}}{Reactions are dropped that have a range of consistency scores narrow than this threshold.}

\item{\code{cluster_strength}}{A number between 0 and 1, specifying the aggressiveness with which to cluster similar reactions together into metareactions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CompassSettings$new()}}
\item \href{#method-print}{\code{CompassSettings$print()}}
\item \href{#method-repr}{\code{CompassSettings$repr()}}
\item \href{#method-clone}{\code{CompassSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassSettings$new(
  ...,
  metabolic_model_directory = system.file("extdata", "RECON2", package = "compassR",
    mustWork = TRUE),
  gene_metadata_file = "gene_metadata.csv",
  metabolite_metadata_file = "metabolite_metadata.csv",
  reaction_metadata_file = "reaction_metadata.csv",
  user_data_directory,
  cell_metadata_file = "cell_metadata.csv",
  compass_scores_file = "reactions.tsv",
  linear_gene_expression_matrix_file = "linear_gene_expression_matrix.tsv",
  cell_id_col_name,
  gene_id_col_name,
  reaction_direction_separator = "_",
  reaction_directions = c("pos", "neg"),
  min_reaction_consistency = 1e-04,
  min_reaction_range = 1e-08,
  cluster_strength = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}

\item{\code{metabolic_model_directory}}{The path to the directory containing the specifications of your metabolic model.}

\item{\code{gene_metadata_file}}{A path to a file in the metabolic_model_directory containing tabular gene metadata. Each row should represent a single gene. The columns are up to you, so long as one of them provides a unique identifier for each gene.}

\item{\code{metabolite_metadata_file}}{A path to a file in the metabolic_model_directory containing tabular metabolite metadata. Each row should represent a single metabolite. The columns are up to you, so long as one of them provides a unique identifier for each metabolite.}

\item{\code{reaction_metadata_file}}{A path to a file in the metabolic_model_directory containing tabular reaction metadata. Each row should represent a single reaction. The columns are up to you, so long as one of them provides a unique identifier for each reaction.}

\item{\code{user_data_directory}}{The path to the directory containing the data specific to the analysis you hope to conduct.}

\item{\code{cell_metadata_file}}{A path to a file in the user_data_directory containing tabular cell metadata. Each row should represent a single cell. The columns are up to you, so long as one of them provides a unique identifier for each cell.}

\item{\code{compass_scores_file}}{A path to a file in the user_data_directory containing the raw reaction consistencies matrix. (This is the output of the COMPASS algorithm.)}

\item{\code{linear_gene_expression_matrix_file}}{A path to a file in the user_data_directory containing the linear gene expression matrix. (This is the input to the COMPASS algorithm.)}

\item{\code{cell_id_col_name}}{The name of the column that uniquely identifies each cell in the cell metadata file.}

\item{\code{gene_id_col_name}}{The name of the column that uniquely identifies each gene in the gene metadata file.}

\item{\code{reaction_direction_separator}}{It is assumed that reaction IDs take the form {unique id}{separator}{one of N annotations}, where the separator is specified by this length-1 character vector, interpreted as a regular expression.}

\item{\code{reaction_directions}}{It is assumed that reaction IDs take the form {unique id}{separator}{one of N annotations}, where the annotations are specified by this length-N character vector, interpreted as regular expressions.}

\item{\code{min_reaction_consistency}}{Reactions are dropped that have consistency scores below this threshold.}

\item{\code{min_reaction_range}}{Reactions are dropped that have a range of consistency scores narrow than this threshold.}

\item{\code{cluster_strength}}{A number between 0 and 1, specifying the aggressiveness with which to cluster similar reactions together into metareactions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints a human-readable representation of this CompassSettings instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassSettings$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\subsection{Method \code{repr()}}{
Returns a human-readable representation of this CompassSettings instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassSettings$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
