% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{genome_track_map}
\alias{genome_track_map}
\title{Function to make the genome track map with combined expression and group information}
\usage{
genome_track_map(
  link_df,
  group_df,
  gene,
  expr_df,
  output_file = "./test.pdf",
  assembly = "hg38",
  bm_mirror = "useast",
  width = 12,
  height = 6,
  ...
)
}
\arguments{
\item{link_df}{data of linkage data frame}

\item{group_df}{data frame of group information}

\item{gene}{gene to plot}

\item{expr_df}{list of gene expression information}

\item{output_file}{output plot file name}

\item{assembly}{bm_mirror useast, asia, www}

\item{width}{plot width}

\item{height}{plot height}

\item{...}{additional parameters passed to plots combind_plots}
}
\value{
plot object
}
\description{
Function to make the genome track map with combined expression and group information
}
\examples{
NA
}
