% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metareaction_consistencies.R
\name{get_annotated_reactions}
\alias{get_annotated_reactions}
\title{Title}
\usage{
get_annotated_reactions(reaction_consistencies, separator, annotations)
}
\arguments{
\item{reaction_consistencies}{the first param}
}
\value{
the output
}
\description{
Description
}
