% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster.R
\name{get_cluster}
\alias{get_cluster}
\title{get_cluster}
\usage{
get_cluster(df, k, method = "average", onlyMembership = TRUE)
}
\arguments{
\item{df}{a matrix or data.frame with coordinates.}

\item{k}{the desired number of clusters.}

\item{method}{the agglomeration method to be used.  This should
    be (an unambiguous abbreviation of) one of
    \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"},
    \code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA),
    \code{"median"} (= WPGMC) or \code{"centroid"} (= UPGMC).}

\item{onlyMembership}{logical. Return only the class membership or the cluster
object.}
}
\value{
Vector of cluster membership or a list with the cluster object and the
distance matrix.
}
\description{
Given a matrix or data.frame with point coordinates and the number of clusters to
get returns the cluster membership of each point or the cluster object with the
distance matrix.
}
\examples{
set.seed(123)
mat <- matrix(c(
    rnorm(n = 100, mean = 10), 
    rnorm(n = 100, mean = 20),
    rnorm(n = 100, mean = 1), 
    rnorm(n = 100, mean = 2)), ncol = 2)
get_cluster(mat, 2)
}
