% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metareaction_consistencies.R
\name{get_metareaction_consistencies}
\alias{get_metareaction_consistencies}
\title{Title}
\usage{
get_metareaction_consistencies(reaction_consistencies, cluster_strength = 0.1)
}
\arguments{
\item{reaction_consistencies}{the first param}

\item{cluster_strength}{the second param}
}
\value{
the output
}
\description{
Description
}
