% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metareaction_consistencies.R
\name{get_metareactions}
\alias{get_metareactions}
\title{Title}
\usage{
get_metareactions(reaction_consistencies, cluster_strength)
}
\arguments{
\item{reaction_consistencies}{the first param}

\item{cluster_strength}{the second param}
}
\value{
a data.frame with columns reaction_name <chr> and metareaction_id <int>.
}
\description{
Description
}
