% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster.R
\name{get_optimal_clusters}
\alias{get_optimal_clusters}
\title{get_optimal_clusters}
\usage{
get_optimal_clusters(
  UMAPcoords,
  k.values = 2:15,
  method = "average",
  plotIt = TRUE
)
}
\arguments{
\item{UMAPcoords}{a matrix or data.frame with the dimensional reduction coordinates.}

\item{k.values}{a vector for the number of clusters to test.}

\item{method}{the agglomeration method to be used.  This should
    be (an unambiguous abbreviation of) one of
    \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"},
    \code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA),
    \code{"median"} (= WPGMC) or \code{"centroid"} (= UPGMC).}

\item{plotIt}{logical. plot or don't plot the output.}
}
\value{
An optional graphical representation of the average silhouette index 
for several clustering configurations and the optimal number of clusters.
}
\description{
Compute the optimal number of clusters for a given set of UMAP coordinates.
}
\examples{
set.seed(123)
mat <- matrix(c(
    rnorm(n = 100, mean = 10), 
    rnorm(n = 100, mean = 20),
    rnorm(n = 100, mean = 1), 
    rnorm(n = 100, mean = 2)), ncol = 2)
get_optimal_clusters(mat)
}
