% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_umap_components.R
\name{get_umap_components}
\alias{get_umap_components}
\title{Title}
\usage{
get_umap_components(consistencies_matrix, ..., num_components = 2)
}
\arguments{
\item{consistencies_matrix}{A param.}

\item{...}{A param.}

\item{num_components}{A param.}
}
\value{
An output.
}
\description{
Description.
}
