% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_peak_group}
\alias{make_peak_group}
\title{separate peaks based on their CRE-gene linkages on diferent groups}
\usage{
make_peak_group(link_gene_df, peaks, df_group)
}
\arguments{
\item{link_gene_df}{CRE-gene link data frame}

\item{peaks}{coordinate of peaks}

\item{df_group}{group information of each sample}
}
\value{
data frame of the peak proportion in different groups
}
\description{
separate peaks based on their CRE-gene linkages on diferent groups
}
\examples{
NA
}
