% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicModel.R
\name{MetabolicModel}
\alias{MetabolicModel}
\title{Title}
\description{
Description.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gene_metadata}}{A field.}

\item{\code{metabolite_metadata}}{A field.}

\item{\code{reaction_metadata}}{A field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetabolicModel$new()}}
\item \href{#method-print}{\code{MetabolicModel$print()}}
\item \href{#method-clone}{\code{MetabolicModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolicModel$new(
  metabolic_model_directory,
  ...,
  gene_metadata_file = "gene_metadata.csv",
  metabolite_metadata_file = "metabolite_metadata.csv",
  reaction_metadata_file = "reaction_metadata.csv"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metabolic_model_directory}}{A param.}

\item{\code{...}}{A param.}

\item{\code{gene_metadata_file}}{A param.}

\item{\code{metabolite_metadata_file}}{A param.}

\item{\code{reaction_metadata_file}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolicModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolicModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
