% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{Metadata}
\alias{Metadata}
\title{Title}
\description{
Description.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gene_metadata}}{A field.}

\item{\code{metabolite_metadata}}{A field.}

\item{\code{reaction_metadata}}{A field.}

\item{\code{compass_metadata}}{A field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Metadata$new()}}
\item \href{#method-print}{\code{Metadata$print()}}
\item \href{#method-repr}{\code{Metadata$repr()}}
\item \href{#method-clone}{\code{Metadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$new(
  gene_metadata,
  metabolite_metadata,
  reaction_metadata,
  compass_metadata
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene_metadata}}{A param.}

\item{\code{metabolite_metadata}}{A param.}

\item{\code{reaction_metadata}}{A param.}

\item{\code{compass_metadata}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\subsection{Method \code{repr()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
