% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_annotation_bar}
\alias{plot_annotation_bar}
\title{Function to plot annotation bar for each sample}
\usage{
plot_annotation_bar(df_group, sample_order = df_group$Sample, ...)
}
\arguments{
\item{df_group}{annotation dataframe with two columns: Sample and Group}

\item{sample_order}{the vector indicate sample order of all samples}

\item{...}{parameters passed to ggplot2::theme}
}
\value{
ggplot2 object
}
\description{
Function to plot annotation bar for each sample
}
\examples{
NA

}
