% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_genome_track}
\alias{plot_genome_track}
\title{functions to plot out the genome track}
\usage{
plot_genome_track(
  df,
  gene,
  assembly,
  sample_ids,
  min_x = NULL,
  max_x = NULL,
  bm_host = "https://nov2020.archive.ensembl.org",
  bm_mirror = NULL
)
}
\arguments{
\item{df}{gene-cre linkage dataframe}

\item{gene}{gene to plot}

\item{assembly}{either mm10 or hg38}

\item{sample_ids}{samples that want to plot}

\item{min_x}{setting the minimum range}

\item{max_x}{setting the maximum range}

\item{bm_host}{biomart host site}

\item{bm_mirror}{biomart useat, asia, www}
}
\value{
list of plots for genome track
}
\description{
functions to plot out the genome track
}
\examples{
NA
}
