% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_giggle}
\alias{plot_giggle}
\title{Plot motif or tf binding information}
\usage{
plot_giggle(motif_df, top_n = 15, title = "")
}
\arguments{
\item{motif_df}{data frame of motif or tf binding information}

\item{top_n}{top n sets to be showed}

\item{title}{title of the plot}
}
\value{
ggplot2 object of motif plot
}
\description{
Plot motif or tf binding information
}
\examples{
NA
}
