% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{plot_metareaction}
\alias{plot_metareaction}
\title{plot_metareactions}
\usage{
plot_metareaction(met_id, wilcoxon_results_with_metadata)
}
\arguments{
\item{met_id}{a vector containing the metareaction names}

\item{wilcoxon_results_with_metadata}{the object created by the 
compare_clusters() function}
}
\value{
an alluvium plot.
}
\description{
Given the wilcoxon results generated by the compare_clusters() function and
the name of some metareactions, the alluvium plot of their metabolism and 
subsystems is generated.
}
