% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_exprssion}
\alias{query_exprssion}
\title{Get expression of each sample}
\usage{
query_exprssion(assembly, gene, celltype = FALSE)
}
\arguments{
\item{assembly}{hg38 or mm10}

\item{gene}{gene symbol}

\item{celltype}{TRUE or FALSE to indicate if this is cell type specific expression}
}
\value{
data frame of expression
}
\description{
Get expression of each sample
}
\examples{
NA
}
