% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_linkage}
\alias{query_linkage}
\title{query linkage of a gene or a genome region}
\usage{
query_linkage(assembly, gene = "", coord = "", celltype = FALSE)
}
\arguments{
\item{assembly}{hg38 or mm10}

\item{gene}{gene symbol}

\item{coord}{genome coordinate in the form of chr-start-end}

\item{celltype}{TRUE or FALSE to indicate if this is cell type specific linkage}
}
\value{
a list with linkage and corresponding sample information
}
\description{
query linkage of a gene or a genome region
}
\examples{
NA
}
