% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_meta}
\alias{query_meta}
\title{query metadata of samples}
\usage{
query_meta(api = "http://scomicsdb.com/metaapi/")
}
\arguments{
\item{api}{website api}
}
\value{
data frame of sample information
}
\description{
query metadata of samples
}
\examples{
NA
}
