% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quick_example_compassr}
\alias{quick_example_compassr}
\title{wrap up function to generate results of examples quickly}
\usage{
quick_example_compassr(gene, output = "./")
}
\arguments{
\item{gene}{gene involved in each example}

\item{output}{output folder for download files}
}
\value{
figures for example results
}
\description{
wrap up function to generate results of examples quickly
}
\examples{
NA
}
