% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{tf_binding}
\alias{tf_binding}
\title{Obtain TF binding information of genome region(s)}
\usage{
tf_binding(assembly, coords)
}
\arguments{
\item{assembly}{hg38/mm10}

\item{coords}{genome coordinate in the form of chr-start-end}
}
\value{
a dataframe of transcription factor binding
}
\description{
Obtain TF binding information of genome region(s)
}
\examples{
NA
}
