% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyzer.R
\name{Analyzer}
\alias{Analyzer}
\title{Title}
\description{
Description.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reaction_consistencies}}{A field.}

\item{\code{metareaction_consistencies}}{A field.}

\item{\code{gene_expression_statistics}}{A field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Analyzer$new()}}
\item \href{#method-clone}{\code{Analyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyzer$new(
  ...,
  metabolic_model,
  reaction_consistencies_file,
  linear_gene_expression_matrix_file,
  min_reaction_consistency = 1e-04,
  min_reaction_range = 1e-08,
  cluster_strength = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}

\item{\code{metabolic_model}}{A param.}

\item{\code{reaction_consistencies_file}}{A param.}

\item{\code{linear_gene_expression_matrix_file}}{A param.}

\item{\code{min_reaction_consistency}}{A param.}

\item{\code{min_reaction_range}}{A param.}

\item{\code{cluster_strength}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
