% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompassAnalyzer.R
\name{CompassAnalyzer}
\alias{CompassAnalyzer}
\title{Title}
\description{
Description.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{settings}}{A field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CompassAnalyzer$new()}}
\item \href{#method-print}{\code{CompassAnalyzer$print()}}
\item \href{#method-repr}{\code{CompassAnalyzer$repr()}}
\item \href{#method-get_umap_components}{\code{CompassAnalyzer$get_umap_components()}}
\item \href{#method-conduct_wilcoxon_test}{\code{CompassAnalyzer$conduct_wilcoxon_test()}}
\item \href{#method-clone}{\code{CompassAnalyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassAnalyzer$new(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassAnalyzer$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\subsection{Method \code{repr()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassAnalyzer$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_umap_components"></a>}}
\subsection{Method \code{get_umap_components()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassAnalyzer$get_umap_components(
  consistencies_matrix,
  ...,
  num_components = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{consistencies_matrix}}{A param.}

\item{\code{...}}{A param.}

\item{\code{num_components}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conduct_wilcoxon_test"></a>}}
\subsection{Method \code{conduct_wilcoxon_test()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassAnalyzer$conduct_wilcoxon_test(
  consistencies_matrix,
  group_A_cell_ids,
  group_B_cell_ids,
  ...,
  reaction_or_metareaction_id_col_name = "metareaction_id"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{consistencies_matrix}}{A param.}

\item{\code{group_A_cell_ids}}{A param.}

\item{\code{group_B_cell_ids}}{A param.}

\item{\code{...}}{A param.}

\item{\code{reaction_or_metareaction_id_col_name}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
