% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompassSettings.R
\name{CompassSettings}
\alias{CompassSettings}
\title{Title}
\description{
Description.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metabolic_model_directory}}{A field.}

\item{\code{gene_metadata_file}}{A field.}

\item{\code{metabolite_metadata_file}}{A field.}

\item{\code{reaction_metadata_file}}{A field.}

\item{\code{user_data_directory}}{A field.}

\item{\code{cell_metadata_file}}{A field.}

\item{\code{reaction_consistencies_file}}{A field.}

\item{\code{linear_gene_expression_matrix_file}}{A field.}

\item{\code{gene_id_col_name}}{A field.}

\item{\code{cell_id_col_name}}{A field.}

\item{\code{reaction_direction_separator}}{A field.}

\item{\code{reaction_directions}}{A field.}

\item{\code{min_reaction_consistency}}{A field.}

\item{\code{min_reaction_range}}{A field.}

\item{\code{cluster_strength}}{A field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CompassSettings$new()}}
\item \href{#method-print}{\code{CompassSettings$print()}}
\item \href{#method-repr}{\code{CompassSettings$repr()}}
\item \href{#method-clone}{\code{CompassSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassSettings$new(
  ...,
  metabolic_model_directory,
  gene_metadata_file,
  metabolite_metadata_file,
  reaction_metadata_file,
  user_data_directory,
  cell_metadata_file,
  reaction_consistencies_file,
  linear_gene_expression_matrix_file,
  gene_id_col_name,
  cell_id_col_name,
  reaction_direction_separator = "_",
  reaction_directions = c("pos", "neg"),
  min_reaction_consistency = 1e-04,
  min_reaction_range = 1e-08,
  cluster_strength = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}

\item{\code{metabolic_model_directory}}{A param.}

\item{\code{gene_metadata_file}}{A param.}

\item{\code{metabolite_metadata_file}}{A param.}

\item{\code{reaction_metadata_file}}{A param.}

\item{\code{user_data_directory}}{A param.}

\item{\code{cell_metadata_file}}{A param.}

\item{\code{reaction_consistencies_file}}{A param.}

\item{\code{linear_gene_expression_matrix_file}}{A param.}

\item{\code{gene_id_col_name}}{A param.}

\item{\code{cell_id_col_name}}{A param.}

\item{\code{reaction_direction_separator}}{A param.}

\item{\code{reaction_directions}}{A param.}

\item{\code{min_reaction_consistency}}{A param.}

\item{\code{min_reaction_range}}{A param.}

\item{\code{cluster_strength}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassSettings$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\subsection{Method \code{repr()}}{
Description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassSettings$repr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A param.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompassSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
